+++ src/config.rs
@@ -22,9 +22,16 @@ impl Manifest {
         use std::fs::File;
         let file = File::open(path)
             .map_err(|cause| Error::chain(format!("Cannot open manifest {:?}!", path), cause))?;
-        let manifest = serde_yaml::from_reader(file).map_err(|cause| {
-            Error::chain(format!("Syntax error in manifest {:?}!", path), cause)
-        })?;
+        let manifest = {
+            let mut manifest: Manifest = serde_yaml::from_reader(file).map_err(|cause| {
+                Error::chain(format!("Syntax error in manifest {:?}!", path), cause)
+            })?;
+            // Add path of Manifest to inputs of each Artifact.
+            for (_, artifact) in &mut manifest.artifacts {
+                artifact.inputs.push(path.to_path_buf())
+            };
+            manifest
+        };
         Ok(manifest)
     }
 }